Shuffle the usr/ directory around.
Previously, usr/ contained code specific to the php cart. Now it
also contains the dsp-core code (1.9.2, as of this commit). usr/php/
contains everything that usr/ previously contained; usr/dsp-core contains
the dsp-core code, in versioned release directories.

The paths used in the scripts in usr/php/* are also changed to point
to the correct location (in most cases, it means changing
.../usr/foo to .../usr/php/foo, but with some subtleties).